Dev to master

+++ app/src/main/java/com/pararam2006/coffeeapp/domain/dto/LocationsDto.kt
@@ -7,6 +7,6 @@ data class LocationsDto(
 )
 
 data class PointDto(
-    val latitude: Int,
-    val longitude: Int,
+    val latitude: Double,
+    val longitude: Double
 )
